feat: implement loan service with creation and repayment logic
Implement complete loan management system with scheduled repayments:

- Add Loan model with outstanding_amount casting
- Add ScheduledRepayment model with amount casting
- Add ReceivedRepayment model for tracking payments
- Implement LoanService.createLoan() with scheduled repayments distribution
  * Generates 3 scheduled repayments by default
  * Distributes loan amount fairly across repayments
- Implement LoanService.repayLoan() with dual-strategy logic
  * Strategy 1: Normalize scheduled repayment amounts when pre-existing REPAID items exist
  * Strategy 2: Reassign due_dates from REPAID repayments to DUE repayments
- Add database migrations for loans, scheduled_repayments, received_repayments
- Add custom ScheduledRepaymentsRelation with sum override
- Add LoanPolicy for authorization checks

Tests: 4/4 PASSING
- Service can create loan for a customer
- Service can repay a scheduled repayment
- Service can repay a scheduled repayment consecutively
- Service can repay multiple scheduled repayments

+++ database/migrations/2020_01_28_100000_create_loans_table.php
@@ -15,10 +15,11 @@ class CreateLoansTable extends Migration
     {
         Schema::create('loans', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
-            $table->integer('amount');
+            $table->unsignedBigInteger('user_id');
+            
+            $table->decimal('amount', 15, 2);
             $table->integer('terms');
-            $table->integer('outstanding_amount');
+            $table->decimal('outstanding_amount', 15, 2);
             $table->string('currency_code');
             $table->date('processed_at');
             $table->string('status');
